feat: support yurtadm local mode for joining IDC node in K8s-on-K8s cluster (#2453)

+++ pkg/yurtadm/cmd/join/phases/prepare.go
@@ -33,9 +33,11 @@ import (
 // RunPrepare executes the node initialization process.
 func RunPrepare(data joindata.YurtJoinData) error {
 	// cleanup at first
-	staticPodsPath := filepath.Join(constants.KubeletConfigureDir, constants.ManifestsSubDirName)
-	if err := os.RemoveAll(staticPodsPath); err != nil {
-		klog.Warningf("remove %s: %v", staticPodsPath, err)
+	if data.NodeRegistration().WorkingMode != constants.LocalNode {
+		staticPodsPath := filepath.Join(constants.KubeletConfigureDir, constants.ManifestsSubDirName)
+		if err := os.RemoveAll(staticPodsPath); err != nil {
+			klog.Warningf("remove %s: %v", staticPodsPath, err)
+		}
 	}
 
 	if err := system.SetIpv4Forward(); err != nil {
@@ -53,8 +55,10 @@ func RunPrepare(data joindata.YurtJoinData) error {
 	if err := yurtadmutil.CheckAndInstallKubeadm(data.KubernetesResourceServer(), data.KubernetesVersion()); err != nil {
 		return err
 	}
-	if err := yurtadmutil.CheckAndInstallKubernetesCni(data.ReuseCNIBin()); err != nil {
-		return err
+	if data.NodeRegistration().WorkingMode != constants.LocalNode {
+		if err := yurtadmutil.CheckAndInstallKubernetesCni(data.ReuseCNIBin()); err != nil {
+			return err
+		}
 	}
 	if err := yurtadmutil.SetKubeletService(); err != nil {
 		return err
@@ -62,22 +66,25 @@ func RunPrepare(data joindata.YurtJoinData) error {
 	if err := yurtadmutil.EnableKubeletService(); err != nil {
 		return err
 	}
-	if err := yurtadmutil.SetKubeletUnitConfig(); err != nil {
+	if err := yurtadmutil.SetKubeletUnitConfig(data); err != nil {
 		return err
 	}
-	if err := yurtadmutil.SetKubeletConfigForNode(); err != nil {
-		return err
-	}
-	if err := yurthub.SetHubBootstrapConfig(data.ServerAddr(), data.JoinToken(), data.CaCertHashes()); err != nil {
-		return err
-	}
-	if err := yurthub.AddYurthubStaticYaml(data, constants.StaticPodPath); err != nil {
-		return err
-	}
-	if len(data.StaticPodTemplateList()) != 0 {
-		// deploy user specified static pods
-		if err := edgenode.DeployStaticYaml(data.StaticPodManifestList(), data.StaticPodTemplateList(), constants.StaticPodPath); err != nil {
+	if data.NodeRegistration().WorkingMode != constants.LocalNode {
+		// the following steps are only for edge/cloud node
+		if err := yurtadmutil.SetKubeletConfigForNode(); err != nil {
 			return err
+		}
+		if err := yurthub.SetHubBootstrapConfig(data.ServerAddr(), data.JoinToken(), data.CaCertHashes()); err != nil {
+			return err
+		}
+		if err := yurthub.AddYurthubStaticYaml(data, constants.StaticPodPath); err != nil {
+			return err
+		}
+		if len(data.StaticPodTemplateList()) != 0 {
+			// deploy user specified static pods
+			if err := edgenode.DeployStaticYaml(data.StaticPodManifestList(), data.StaticPodTemplateList(), constants.StaticPodPath); err != nil {
+				return err
+			}
 		}
 	}
 	if err := yurtadmutil.SetDiscoveryConfig(data); err != nil {
